api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10